+++ packages/react-api/src/types.d.ts
@@ -43,10 +43,8 @@ export type SourceProps = {
   type: MapTypesType['QUERY'] | MapTypesType['TABLE'] | MapTypesType['TILESET'];
   connection: string;
   geoColumn?: string;
-  dataResolution?: number;
   spatialDataType?: string;
   spatialDataColumn?: string;
-  spatialFiltersResolution?: number;
   aggregationExp?: string;
   aggregationResLevel?: number;
   credentials?: Credentials;