+++ packages/react-widgets/src/index.js
@@ -31,4 +31,3 @@ export {
   sourceAndFiltersToSQL as _sourceAndFiltersToSQL,
   getSqlEscapedSource as _getSqlEscapedSource
 } from './models/utils';
-export { getSpatialFiltersResolution as _getSpatialFiltersResolution } from './models/spatialFiltersResolution';